temp

+++ drv_stm32/FW_Src/Inc/drv_adc_core.h
@@ -8,7 +8,8 @@
 void drv_adc_init (void);
 void drv_adc_start_of_conversion (uint8_t adc_id);
 void drv_adc_calibrate (uint8_t adc_id);
-float drv_adc_get_value (uint8_t adc_id);
+uint16_t drv_adc_get_value (uint8_t adc_id);
+
 
 #endif /* ENABLE_ADC */
 #endif /* DRV_ADC_CORE_H_ */
